CPP06 - ex00 - char

+++ cpp-06/ex00/src/ScalarPrinter.cpp
@@ -6,7 +6,7 @@
 /*   By: mfidimal <mfidimal@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/29 04:38:50 by mfidimal          #+#    #+#             */
-/*   Updated: 2025/12/29 05:19:40 by mfidimal         ###   ########.fr       */
+/*   Updated: 2025/12/29 05:39:53 by mfidimal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,9 @@
 
 void ScalarPrinter::printCharBase(char const &chr) {
   std::cout << "char: " << chr << std::endl
-            << "int: " << chr << std::endl
-            << "float: " << chr << ".0f" << std::endl
-            << "double: " << chr << ".0" << std::endl;
+            << "int: " << static_cast<int>(chr) << std::endl
+            << "float: " << static_cast<float>(chr) << ".0f" << std::endl
+            << "double: " << static_cast<double>(chr) << ".0" << std::endl;
 }
 
 void ScalarPrinter::printInvalidBase() {
